Add different onXX handling stuff because it was fun! :)
when we added the 'value' attribute to the input and set the value
to be our state, so we needed to add the 'onChange' too to be
able to control it. Define what should happen when we have a value
on the input. Otherwise React(?) will block the functionality
for us, since the 'value' attribute is set to our State ('newNote')